requirements.txt
Adafruit-Blinka==8.60.3
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-framebuf==1.6.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-ssd1306==2.12.21
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.80.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
cbor2==5.6.5
lgpio==0.2.2.0
luma.core==2.4.2
luma.oled==3.14.0
pillow==11.2.1
pyftdi==0.56.0
pyserial==3.5
python-rtmidi==1.5.8
pyusb==1.3.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
smbus2==0.5.0
spidev==3.7
sysv_ipc==1.1.0
typing_extensions==4.14.0
